refactor: move cabin interfaces to separate file
